+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### v1.1.47 (2014-08-28)
+* * *
+* Customer id can be passed to the checkout new subscription operation.
+
+* Added support for affiliate integration to accept affiliate token and the ip address from where the subscription was created. See https://apidocs.chargebee.com/docs/api/subscriptions#create_a_subscription.
+
 ### v1.1.46 (2014-08-13)
 * * *
 Added properties: